refactor category destroying notification

+++ app/controllers/web/admin/categories_controller.rb
@@ -34,10 +34,15 @@ class Web::Admin::CategoriesController < Web::Admin::ApplicationController
   end
 
   def destroy
+    if @category.bulletins.count.positive?
+      redirect_to admin_categories_path, alert: t('.cant_delete_category')
+      return
+    end
+
     if @category.destroy
-      redirect_to admin_categories_path
+      redirect_to admin_categories_path, notice: t('.success')
     else
-      redirect_to admin_categories_path, { alert: @category.errors.full_messages.to_sentence }
+      redirect_to admin_categories_path, alert: @category.errors.full_messages.to_sentence
     end
   end
